+++ R/options.R
@@ -60,7 +60,7 @@ find_args <- function () {
 #'        Note: RStudio only supports 8 bit ANSI output (24bit ANSI is
 #'        rendered invisibly in Rstudio).  For 24bit colour output, try R in the terminal
 #'        e.g. 'iTerm' on OSX.
-#' @param text_mode How to handle textif no text colour has been
+#' @param text_mode How to handle text if no text colour has been
 #'        explicitly specified by the user.
 #'        \describe{
 #'        \item{contrast}{(default) automatically select a contrasting colour for enhanced readability.}